+++ src/components/EmptyDashboard/index.js
@@ -3,19 +3,26 @@ import PropTypes from 'prop-types';
 
 const EmptyDashboard = ({
   button,
+  imageHeight,
   imagePath,
+  imageWidth,
   headline,
   description,
   descriptionMaxWidth,
+  showButtonBelow,
   testSection,
 }) => {
   return (
     <div
       data-oui-component={ true }
-      className='flex overflow-y--scroll flex--1'
-      data-test-section='layers-empty-state'
-      style={{ width: '750px', margin: 'auto' }}>
-      <div className='anchor--middle height--1-1 flex flex--column'>
+      className='flex flex--column overflow-y--scroll flex--1'
+      data-test-section='layers-empty-state'>
+      { button && !showButtonBelow && (
+        <div className="flex push-quad--sides oui-empty-dashboard__button-row flex-justified--end">
+          { button }
+        </div>
+      ) }
+      <div className='anchor--middle height--1-1 flex flex--row flex-align--center'>
 
         <div
           data-oui-component={ true }
@@ -23,13 +30,16 @@ const EmptyDashboard = ({
           data-test-section={ testSection }>
 
           { imagePath && (
-            <div
-              className="muted push-double--bottom"
-              style={{ width: '550px', margin: 'auto' }}>
-              <object
-                data={ imagePath }
+            <div className="push-double--bottom">
+              <img
+                src={ imagePath }
                 className="svg--non-scaling-stroke display--inline"
-                style={{ maxWidth: '450px', maxHeight: '290px' }}
+                style={{
+                  maxWidth: '450px',
+                  maxHeight: '290px',
+                  height: imageHeight,
+                  width: imageWidth,
+                }}
                 data-test-section={ testSection && `${testSection}-image` }
                 alt=""
               />
@@ -53,7 +63,7 @@ const EmptyDashboard = ({
             </div>
           ) }
 
-          { button && (
+          { button && showButtonBelow && (
             <div className="push-double--top">
               { button }
             </div>
@@ -77,14 +87,26 @@ EmptyDashboard.propTypes = {
   ]),
   /** Short text about the empty state */
   headline: PropTypes.string.isRequired,
+  /** String to set the max image height */
+  imageHeight: PropTypes.string,
   /** Path to an image representing the empty state */
   imagePath: PropTypes.string,
+  /** String to set the max image width */
+  imageWidth: PropTypes.string,
+  /**
+   * Boolean used to show the button below the image
+   * instead of top right corner of the empty state
+   */
+  showButtonBelow: PropTypes.bool,
   /** Identifier used to create data-test-section attributes for testing */
   testSection: PropTypes.string,
 };
 
 EmptyDashboard.getDefaultProps = {
   descriptionMaxWidth: '450px',
+  imageHeight: 'inherit',
+  imageWidth: 'inherit',
+  showButtonBelow: false,
   testSection: 'empty-hello',
 };